backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Phase II Todo App - FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
